refactor(iota-core, iota-e2e-tests, iota-tool): [P-COOL] split iota-core authority_client and authority_server into per-service modules (#10954)
Splits the monolithic authority_server.rs and authority_client.rs into
module directories with separate files per gRPC service (validator,
validator_peer, validator_v2). Shared infrastructure (server handle,
metrics, traffic control, client struct, connection helpers) lives in
mod.rs, while each service's trait impl gets its own file.

closes #10945.

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [ ] Patch-specific tests (correctness, functionality coverage)
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have checked that new and existing unit tests pass locally with
my changes
