+++ crates/iota-core/src/authority_client/validator.rs
@@ -3,21 +3,12 @@
 // Modifications Copyright (c) 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use std::{collections::BTreeMap, net::SocketAddr, time::Duration};
+use std::net::SocketAddr;
 
-use anyhow::anyhow;
 use async_trait::async_trait;
-use iota_network::{
-    api::ValidatorClient,
-    tonic,
-    tonic::{metadata::KeyAndValueRef, transport::Channel},
-};
-use iota_network_stack::config::Config;
+use iota_network::tonic;
 use iota_types::{
-    base_types::AuthorityName,
-    committee::CommitteeWithNetworkMetadata,
-    crypto::NetworkPublicKey,
-    error::{IotaError, IotaResult},
+    error::IotaError,
     iota_system_state::IotaSystemState,
     messages_checkpoint::{CheckpointRequest, CheckpointResponse},
     messages_grpc::{
@@ -29,16 +20,14 @@ use iota_types::{
         TransactionInfoRequest, TransactionInfoResponse, ValidatorHealthRequest,
         ValidatorHealthResponse, WaitForEffectsRequest, WaitForEffectsResponse,
     },
-    multiaddr::Multiaddr,
     transaction::*,
 };
-use tap::TapFallible;
 
-use crate::authority_client::tonic::IntoRequest;
+use crate::authority_client::{NetworkAuthorityClient, insert_metadata, tonic::IntoRequest};
 
 #[async_trait]
-pub trait AuthorityAPI {
-    /// Handles a `Transaction` for this account.
+pub trait ValidatorAPI {
+    /// Handles a `Transaction`.
     async fn handle_transaction(
         &self,
         transaction: Transaction,
@@ -59,19 +48,19 @@ pub trait AuthorityAPI {
         client_addr: Option<SocketAddr>,
     ) -> Result<HandleSoftBundleCertificatesResponseV1, IotaError>;
 
-    /// Handle Object information requests for this account.
+    /// Handle Object information requests.
     async fn handle_object_info_request(
         &self,
         request: ObjectInfoRequest,
     ) -> Result<ObjectInfoResponse, IotaError>;
 
-    /// Handles a `TransactionInfoRequest` for this account.
+    /// Handles a `TransactionInfoRequest`.
     async fn handle_transaction_info_request(
         &self,
         request: TransactionInfoRequest,
     ) -> Result<TransactionInfoResponse, IotaError>;
 
-    /// Handles a `CheckpointRequest` for this account.
+    /// Handles a `CheckpointRequest`.
     async fn handle_checkpoint(
         &self,
         request: CheckpointRequest,
@@ -111,66 +100,9 @@ pub trait AuthorityAPI {
     ) -> Result<ValidatorHealthResponse, IotaError>;
 }
 
-/// A client for the network authority.
-#[derive(Clone)]
-pub struct NetworkAuthorityClient {
-    client: IotaResult<ValidatorClient<Channel>>,
-}
-
-impl NetworkAuthorityClient {
-    pub async fn connect(
-        address: &Multiaddr,
-        tls_target: Option<NetworkPublicKey>,
-    ) -> anyhow::Result<Self> {
-        let tls_config = tls_target.map(|tls_target| {
-            iota_tls::create_rustls_client_config(
-                tls_target,
-                iota_tls::IOTA_VALIDATOR_SERVER_NAME.to_string(),
-                None,
-            )
-        });
-        let channel = iota_network_stack::client::connect(address, tls_config)
-            .await
-            .map_err(|err| anyhow!(err.to_string()))?;
-        Ok(Self::new(channel))
-    }
-
-    pub fn connect_lazy(address: &Multiaddr, tls_target: Option<NetworkPublicKey>) -> Self {
-        let tls_config = tls_target.map(|tls_target| {
-            iota_tls::create_rustls_client_config(
-                tls_target,
-                iota_tls::IOTA_VALIDATOR_SERVER_NAME.to_string(),
-                None,
-            )
-        });
-        let client: IotaResult<_> = iota_network_stack::client::connect_lazy(address, tls_config)
-            .map(ValidatorClient::new)
-            .map_err(|err| err.to_string().into());
-        Self { client }
-    }
-
-    /// Creates a new client with a `transport` channel.
-    pub fn new(channel: Channel) -> Self {
-        Self {
-            client: Ok(ValidatorClient::new(channel)),
-        }
-    }
-
-    /// Creates a new client with a lazy `transport` channel.
-    fn new_lazy(client: IotaResult<Channel>) -> Self {
-        Self {
-            client: client.map(ValidatorClient::new),
-        }
-    }
-
-    fn client(&self) -> IotaResult<ValidatorClient<Channel>> {
-        self.client.clone()
-    }
-}
-
 #[async_trait]
-impl AuthorityAPI for NetworkAuthorityClient {
-    /// Handles a `Transaction` for this account.
+impl ValidatorAPI for NetworkAuthorityClient {
+    /// Handles a `Transaction` .
     async fn handle_transaction(
         &self,
         transaction: Transaction,
@@ -220,7 +152,7 @@ impl AuthorityAPI for NetworkAuthorityClient {
         response.map_err(Into::into)
     }
 
-    /// Handles a `ObjectInfoRequest` for this account.
+    /// Handles a `ObjectInfoRequest` .
     async fn handle_object_info_request(
         &self,
         request: ObjectInfoRequest,
@@ -232,7 +164,7 @@ impl AuthorityAPI for NetworkAuthorityClient {
             .map_err(Into::into)
     }
 
-    /// Handles a `TransactionInfoRequest` for this account.
+    /// Handles a `TransactionInfoRequest` .
     async fn handle_transaction_info_request(
         &self,
         request: TransactionInfoRequest,
@@ -244,7 +176,7 @@ impl AuthorityAPI for NetworkAuthorityClient {
             .map_err(Into::into)
     }
 
-    /// Handles a `CheckpointRequest` for this account.
+    /// Handles a `CheckpointRequest` .
     async fn handle_checkpoint(
         &self,
         request: CheckpointRequest,
@@ -329,76 +261,5 @@ impl AuthorityAPI for NetworkAuthorityClient {
             .map(tonic::Response::into_inner)
             .map_err(IotaError::from)?;
         raw_response.try_into()
-    }
-}
-
-/// Creates authority clients with network configuration.
-pub fn make_network_authority_clients_with_network_config(
-    committee: &CommitteeWithNetworkMetadata,
-    network_config: &Config,
-) -> BTreeMap<AuthorityName, NetworkAuthorityClient> {
-    let mut authority_clients = BTreeMap::new();
-    for (name, (_state, network_metadata)) in committee.validators() {
-        let address = network_metadata
-            .network_address
-            .clone()
-            .rewrite_udp_to_tcp()
-            .rewrite_http_to_https();
-        let tls_config = network_metadata
-            .network_public_key
-            .as_ref()
-            .map(|key| {
-                iota_tls::create_rustls_client_config(
-                    key.clone(),
-                    iota_tls::IOTA_VALIDATOR_SERVER_NAME.to_string(),
-                    None,
-                )
-            })
-            .ok_or(IotaError::from("network public key is not available"));
-        let maybe_channel = tls_config
-            .and_then(|tls_config| {
-                network_config
-                    .connect_lazy(&address, Some(tls_config))
-                    .map_err(|e| e.to_string().into())
-            })
-            .tap_err(|e| {
-                tracing::error!(
-                    address = %address,
-                    name = %name,
-                    "unable to create authority client: {e}"
-                )
-            });
-        let client = NetworkAuthorityClient::new_lazy(maybe_channel);
-        authority_clients.insert(*name, client);
-    }
-    authority_clients
-}
-
-/// Creates authority clients with a timeout configuration.
-pub fn make_authority_clients_with_timeout_config(
-    committee: &CommitteeWithNetworkMetadata,
-    connect_timeout: Duration,
-    request_timeout: Duration,
-) -> BTreeMap<AuthorityName, NetworkAuthorityClient> {
-    let mut network_config = iota_network_stack::config::Config::new();
-    network_config.connect_timeout = Some(connect_timeout);
-    network_config.request_timeout = Some(request_timeout);
-    make_network_authority_clients_with_network_config(committee, &network_config)
-}
-
-fn insert_metadata<T>(request: &mut tonic::Request<T>, client_addr: Option<SocketAddr>) {
-    if let Some(client_addr) = client_addr {
-        let mut metadata = tonic::metadata::MetadataMap::new();
-        metadata.insert("x-forwarded-for", client_addr.to_string().parse().unwrap());
-        metadata
-            .iter()
-            .for_each(|key_and_value| match key_and_value {
-                KeyAndValueRef::Ascii(key, value) => {
-                    request.metadata_mut().insert(key, value.clone());
-                }
-                KeyAndValueRef::Binary(key, value) => {
-                    request.metadata_mut().insert_bin(key, value.clone());
-                }
-            });
     }
 }
