+++ crates/iota-core/src/authority_server/mod.rs
@@ -1,0 +1,280 @@
+// Copyright (c) 2021, Facebook, Inc. and its affiliates
+// Copyright (c) Mysten Labs, Inc.
+// Modifications Copyright (c) 2024 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+mod validator;
+mod validator_peer;
+mod validator_v2;
+
+pub mod metrics;
+#[cfg(test)]
+mod test_server;
+
+use std::{
+    net::{IpAddr, SocketAddr},
+    sync::Arc,
+    time::SystemTime,
+};
+
+use iota_network::{
+    tonic,
+    tonic::metadata::{Ascii, MetadataValue},
+};
+use iota_types::{
+    error::*,
+    traffic_control::{ClientIdSource, PolicyConfig, RemoteFirewallConfig, Weight},
+};
+pub use metrics::ValidatorServiceMetrics;
+#[cfg(test)]
+pub use test_server::{AuthorityServer, AuthorityServerHandle};
+use tonic::transport::server::TcpConnectInfo;
+use tracing::error;
+
+use crate::{
+    authority::AuthorityState,
+    consensus_adapter::ConsensusAdapter,
+    traffic_controller::{
+        TrafficController, metrics::TrafficControllerMetrics, parse_ip, policies::TrafficTally,
+    },
+};
+type WrappedServiceResponse<T> = Result<(tonic::Response<T>, Weight), tonic::Status>;
+
+/// The validator service.
+#[derive(Clone)]
+pub struct ValidatorService {
+    state: Arc<AuthorityState>,
+    consensus_adapter: Arc<ConsensusAdapter>,
+    metrics: Arc<ValidatorServiceMetrics>,
+    traffic_controller: Option<Arc<TrafficController>>,
+    client_id_source: Option<ClientIdSource>,
+}
+
+impl ValidatorService {
+    /// Creates a new `ValidatorService`.
+    pub fn new(
+        state: Arc<AuthorityState>,
+        consensus_adapter: Arc<ConsensusAdapter>,
+        validator_metrics: Arc<ValidatorServiceMetrics>,
+        traffic_controller_metrics: TrafficControllerMetrics,
+        policy_config: Option<PolicyConfig>,
+        firewall_config: Option<RemoteFirewallConfig>,
+    ) -> Self {
+        Self {
+            state,
+            consensus_adapter,
+            metrics: validator_metrics,
+            traffic_controller: policy_config.clone().map(|policy| {
+                Arc::new(TrafficController::init(
+                    policy,
+                    traffic_controller_metrics,
+                    firewall_config,
+                ))
+            }),
+            client_id_source: policy_config.map(|policy| policy.client_id_source),
+        }
+    }
+
+    pub fn new_for_tests(
+        state: Arc<AuthorityState>,
+        consensus_adapter: Arc<ConsensusAdapter>,
+        metrics: Arc<ValidatorServiceMetrics>,
+    ) -> Self {
+        Self {
+            state,
+            consensus_adapter,
+            metrics,
+            traffic_controller: None,
+            client_id_source: None,
+        }
+    }
+
+    /// Returns the validator state.
+    pub fn validator_state(&self) -> &Arc<AuthorityState> {
+        &self.state
+    }
+
+    fn get_client_ip_addr<T>(
+        &self,
+        request: &tonic::Request<T>,
+        source: &ClientIdSource,
+    ) -> Option<IpAddr> {
+        match source {
+            ClientIdSource::SocketAddr => {
+                let socket_addr: Option<SocketAddr> = request.remote_addr();
+
+                // We will hit this case if the IO type used does not
+                // implement Connected or when using a unix domain socket.
+                // TODO: once we have confirmed that no legitimate traffic
+                // is hitting this case, we should reject such requests that
+                // hit this case.
+                if let Some(socket_addr) = socket_addr {
+                    Some(socket_addr.ip())
+                } else {
+                    if cfg!(msim) {
+                        // Ignore the error from simtests.
+                    } else if cfg!(test) {
+                        panic!("Failed to get remote address from request");
+                    } else {
+                        self.metrics.connection_ip_not_found.inc();
+                        error!("Failed to get remote address from request");
+                    }
+                    None
+                }
+            }
+            ClientIdSource::XForwardedFor(num_hops) => {
+                let do_header_parse = |op: &MetadataValue<Ascii>| {
+                    match op.to_str() {
+                        Ok(header_val) => {
+                            let header_contents =
+                                header_val.split(',').map(str::trim).collect::<Vec<_>>();
+                            if *num_hops == 0 {
+                                error!(
+                                    "x-forwarded-for: 0 specified. x-forwarded-for contents: {:?}. Please assign nonzero value for \
+                                    number of hops here, or use `socket-addr` client-id-source type if requests are not being proxied \
+                                    to this node. Skipping traffic controller request handling.",
+                                    header_contents,
+                                );
+                                return None;
+                            }
+                            let contents_len = header_contents.len();
+                            if contents_len < *num_hops {
+                                error!(
+                                    "x-forwarded-for header value of {:?} contains {} values, but {} hops were specified. \
+                                    Expected at least {} values. Please correctly set the `x-forwarded-for` value under \
+                                    `client-id-source` in the node config.",
+                                    header_contents, contents_len, num_hops, contents_len,
+                                );
+                                self.metrics.client_id_source_config_mismatch.inc();
+                                return None;
+                            }
+                            let Some(client_ip) = header_contents.get(contents_len - num_hops)
+                            else {
+                                error!(
+                                    "x-forwarded-for header value of {:?} contains {} values, but {} hops were specified. \
+                                    Expected at least {} values. Skipping traffic controller request handling.",
+                                    header_contents, contents_len, num_hops, contents_len,
+                                );
+                                return None;
+                            };
+                            parse_ip(client_ip).or_else(|| {
+                                self.metrics.forwarded_header_parse_error.inc();
+                                None
+                            })
+                        }
+                        Err(e) => {
+                            // TODO: once we have confirmed that no legitimate traffic
+                            // is hitting this case, we should reject such requests that
+                            // hit this case.
+                            self.metrics.forwarded_header_invalid.inc();
+                            error!("Invalid UTF-8 in x-forwarded-for header: {:?}", e);
+                            None
+                        }
+                    }
+                };
+                if let Some(op) = request.metadata().get("x-forwarded-for") {
+                    do_header_parse(op)
+                } else if let Some(op) = request.metadata().get("X-Forwarded-For") {
+                    do_header_parse(op)
+                } else {
+                    self.metrics.forwarded_header_not_included.inc();
+                    error!(
+                        "x-forwarded-for header not present for request despite node configuring x-forwarded-for tracking type"
+                    );
+                    None
+                }
+            }
+        }
+    }
+
+    async fn handle_traffic_req(&self, client: Option<IpAddr>) -> Result<(), tonic::Status> {
+        if let Some(traffic_controller) = &self.traffic_controller {
+            if !traffic_controller.check(&client, &None).await {
+                // Entity in blocklist
+                Err(tonic::Status::from_error(IotaError::TooManyRequests.into()))
+            } else {
+                Ok(())
+            }
+        } else {
+            Ok(())
+        }
+    }
+
+    fn handle_traffic_resp<T>(
+        &self,
+        client: Option<IpAddr>,
+        wrapped_response: WrappedServiceResponse<T>,
+    ) -> Result<tonic::Response<T>, tonic::Status> {
+        let (error, spam_weight, unwrapped_response) = match wrapped_response {
+            Ok((result, spam_weight)) => (None, spam_weight, Ok(result)),
+            Err(status) => (
+                Some(IotaError::from(status.clone())),
+                Weight::zero(),
+                Err(status),
+            ),
+        };
+
+        if let Some(traffic_controller) = self.traffic_controller.clone() {
+            traffic_controller.tally(TrafficTally {
+                direct: client,
+                through_fullnode: None,
+                error_info: error.map(|e| {
+                    let error_type = String::from(e.as_ref());
+                    let error_weight = normalize(e);
+                    (error_weight, error_type)
+                }),
+                spam_weight,
+                timestamp: SystemTime::now(),
+            })
+        }
+        unwrapped_response
+    }
+}
+
+fn make_tonic_request_for_testing<T>(message: T) -> tonic::Request<T> {
+    // simulate a TCP connection, which would have added extensions to
+    // the request object that would be used downstream
+    let mut request = tonic::Request::new(message);
+    let tcp_connect_info = TcpConnectInfo {
+        local_addr: None,
+        remote_addr: Some(SocketAddr::new([127, 0, 0, 1].into(), 0)),
+    };
+    request.extensions_mut().insert(tcp_connect_info);
+    request
+}
+
+// TODO: refine error matching here
+fn normalize(err: IotaError) -> Weight {
+    match err {
+        IotaError::UserInput {
+            error: UserInputError::IncorrectUserSignature { .. },
+        } => Weight::one(),
+        IotaError::InvalidSignature { .. }
+        | IotaError::SignerSignatureAbsent { .. }
+        | IotaError::SignerSignatureNumberMismatch { .. }
+        | IotaError::IncorrectSigner { .. }
+        | IotaError::UnknownSigner { .. }
+        | IotaError::WrongEpoch { .. } => Weight::one(),
+        _ => Weight::zero(),
+    }
+}
+
+/// Implements generic pre- and post-processing. Since this is on the critical
+/// path, any heavy lifting should be done in a separate non-blocking task
+/// unless it is necessary to override the return value.
+#[macro_export]
+macro_rules! handle_with_decoration {
+    ($self:ident, $func_name:ident, $request:ident) => {{
+        if $self.client_id_source.is_none() {
+            return $self.$func_name($request).await.map(|(result, _)| result);
+        }
+
+        let client = $self.get_client_ip_addr(&$request, $self.client_id_source.as_ref().unwrap());
+
+        // check if either IP is blocked, in which case return early
+        $self.handle_traffic_req(client.clone()).await?;
+
+        // handle traffic tallying
+        let wrapped_response = $self.$func_name($request).await;
+        $self.handle_traffic_resp(client, wrapped_response)
+    }};
+}
