+++ crates/iota-core/src/authority_server/validator.rs
@@ -3,27 +3,16 @@
 // Modifications Copyright (c) 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use std::{
-    io,
-    net::{IpAddr, SocketAddr},
-    sync::Arc,
-    time::SystemTime,
-};
+use std::sync::Arc;
 
 use anyhow::Result;
 use async_trait::async_trait;
-use fastcrypto::traits::KeyPair;
 use futures::future::{Either, join_all};
-use iota_config::local_ip_utils::new_local_tcp_address_for_testing;
 use iota_metrics::spawn_monitored_task;
-use iota_network::{
-    api::{Validator, ValidatorServer},
-    tonic,
-};
-use iota_network_stack::server::IOTA_TLS_SERVER_NAME;
+use iota_network::{api::Validator, tonic};
 use iota_types::{
     effects::{TransactionEffects, TransactionEffectsAPI},
-    error::*,
+    error::{IotaError, UserInputError},
     fp_ensure,
     iota_system_state::IotaSystemState,
     message_envelope::Message,
@@ -37,378 +26,24 @@ use iota_types::{
         ObjectInfoResponse, SubmitCertificateResponse, SubmitTransactionResult, SystemStateRequest,
         TransactionInfoRequest, TransactionInfoResponse,
     },
-    multiaddr::Multiaddr,
-    traffic_control::{ClientIdSource, PolicyConfig, RemoteFirewallConfig, Weight},
+    traffic_control::{ClientIdSource, Weight},
     transaction::*,
 };
 use nonempty::{NonEmpty, nonempty};
-use prometheus::{
-    Histogram, IntCounter, IntCounterVec, Registry, register_histogram_with_registry,
-    register_int_counter_vec_with_registry, register_int_counter_with_registry,
-};
 use tap::TapFallible;
-use tonic::{
-    metadata::{Ascii, MetadataValue},
-    transport::server::TcpConnectInfo,
-};
-use tracing::{Instrument, debug, error, error_span, info, trace_span};
+use tracing::{Instrument, debug, error_span, info, trace_span};
 
 use crate::{
-    authority::{AuthorityState, authority_per_epoch_store::AuthorityPerEpochStore},
-    checkpoints::CheckpointStore,
-    consensus_adapter::{
-        ConnectionMonitorStatusForTests, ConsensusAdapter, ConsensusAdapterMetrics,
-    },
-    mysticeti_adapter::LazyMysticetiClient,
-    traffic_controller::{
-        TrafficController, metrics::TrafficControllerMetrics, parse_ip, policies::TrafficTally,
-    },
+    authority::authority_per_epoch_store::AuthorityPerEpochStore,
+    authority_server::{ValidatorService, WrappedServiceResponse, make_tonic_request_for_testing},
+    handle_with_decoration,
 };
 
 #[cfg(test)]
-#[path = "unit_tests/server_tests.rs"]
+#[path = "../unit_tests/server_tests.rs"]
 mod server_tests;
 
-/// A handle to the authority server.
-pub struct AuthorityServerHandle {
-    server_handle: iota_network_stack::server::Server,
-}
-
-impl AuthorityServerHandle {
-    /// Waits for the server to complete.
-    pub async fn join(self) -> Result<(), io::Error> {
-        self.server_handle.handle().wait_for_shutdown().await;
-        Ok(())
-    }
-
-    /// Kills the server.
-    pub async fn kill(self) -> Result<(), io::Error> {
-        self.server_handle.handle().shutdown().await;
-        Ok(())
-    }
-
-    /// Returns the address of the server.
-    pub fn address(&self) -> &Multiaddr {
-        self.server_handle.local_addr()
-    }
-}
-
-/// An authority server that is used for testing.
-pub struct AuthorityServer {
-    address: Multiaddr,
-    pub state: Arc<AuthorityState>,
-    consensus_adapter: Arc<ConsensusAdapter>,
-    pub metrics: Arc<ValidatorServiceMetrics>,
-}
-
-impl AuthorityServer {
-    /// Creates a new `AuthorityServer` for testing with a consensus adapter.
-    pub fn new_for_test_with_consensus_adapter(
-        state: Arc<AuthorityState>,
-        consensus_adapter: Arc<ConsensusAdapter>,
-    ) -> Self {
-        let address = new_local_tcp_address_for_testing();
-        let metrics = Arc::new(ValidatorServiceMetrics::new_for_tests());
-
-        Self {
-            address,
-            state,
-            consensus_adapter,
-            metrics,
-        }
-    }
-
-    /// Creates a new `AuthorityServer` for testing.
-    pub fn new_for_test(state: Arc<AuthorityState>) -> Self {
-        let consensus_adapter = Arc::new(ConsensusAdapter::new(
-            Arc::new(LazyMysticetiClient::new()),
-            CheckpointStore::new_for_tests(),
-            state.name,
-            Arc::new(ConnectionMonitorStatusForTests {}),
-            100_000,
-            100_000,
-            None,
-            None,
-            ConsensusAdapterMetrics::new_test(),
-        ));
-        Self::new_for_test_with_consensus_adapter(state, consensus_adapter)
-    }
-
-    /// Spawns the server.
-    pub async fn spawn_for_test(self) -> Result<AuthorityServerHandle, io::Error> {
-        let address = self.address.clone();
-        self.spawn_with_bind_address_for_test(address).await
-    }
-
-    /// Spawns the server with a bind address.
-    pub async fn spawn_with_bind_address_for_test(
-        self,
-        address: Multiaddr,
-    ) -> Result<AuthorityServerHandle, io::Error> {
-        let tls_config = iota_tls::create_rustls_server_config(
-            self.state.config.network_key_pair().copy().private(),
-            IOTA_TLS_SERVER_NAME.to_string(),
-        );
-        let server = iota_network_stack::config::Config::new()
-            .server_builder()
-            .add_service(ValidatorServer::new(ValidatorService::new_for_tests(
-                self.state,
-                self.consensus_adapter,
-                self.metrics,
-            )))
-            .bind(&address, Some(tls_config))
-            .await
-            .unwrap();
-        let local_addr = server.local_addr().to_owned();
-        info!("Listening to traffic on {local_addr}");
-        let handle = AuthorityServerHandle {
-            server_handle: server,
-        };
-        Ok(handle)
-    }
-}
-
-/// Metrics for the validator service.
-pub struct ValidatorServiceMetrics {
-    pub signature_errors: IntCounter,
-    pub tx_verification_latency: Histogram,
-    pub cert_verification_latency: Histogram,
-    pub consensus_latency: Histogram,
-    pub handle_transaction_latency: Histogram,
-    pub submit_certificate_consensus_latency: Histogram,
-    pub handle_certificate_consensus_latency: Histogram,
-    pub handle_certificate_non_consensus_latency: Histogram,
-    pub handle_soft_bundle_certificates_consensus_latency: Histogram,
-    pub handle_soft_bundle_certificates_count: Histogram,
-    pub handle_soft_bundle_certificates_size_bytes: Histogram,
-    pub handle_capability_notification_latency: Histogram,
-
-    num_rejected_tx_in_epoch_boundary: IntCounter,
-    num_rejected_cert_in_epoch_boundary: IntCounter,
-    num_rejected_tx_during_overload: IntCounterVec,
-    num_rejected_cert_during_overload: IntCounterVec,
-    num_rejected_capability_notifications_during_overload: IntCounterVec,
-    connection_ip_not_found: IntCounter,
-    forwarded_header_parse_error: IntCounter,
-    forwarded_header_invalid: IntCounter,
-    forwarded_header_not_included: IntCounter,
-    client_id_source_config_mismatch: IntCounter,
-}
-
-impl ValidatorServiceMetrics {
-    /// Creates a new `ValidatorServiceMetrics` with Prometheus registry.
-    pub fn new(registry: &Registry) -> Self {
-        Self {
-            signature_errors: register_int_counter_with_registry!(
-                "total_signature_errors",
-                "Number of transaction signature errors",
-                registry,
-            )
-            .unwrap(),
-            tx_verification_latency: register_histogram_with_registry!(
-                "validator_service_tx_verification_latency",
-                "Latency of verifying a transaction",
-                iota_metrics::SUBSECOND_LATENCY_SEC_BUCKETS.to_vec(),
-                registry,
-            )
-            .unwrap(),
-            cert_verification_latency: register_histogram_with_registry!(
-                "validator_service_cert_verification_latency",
-                "Latency of verifying a certificate",
-                iota_metrics::SUBSECOND_LATENCY_SEC_BUCKETS.to_vec(),
-                registry,
-            )
-            .unwrap(),
-            consensus_latency: register_histogram_with_registry!(
-                "validator_service_consensus_latency",
-                "Time spent between submitting a shared obj txn to consensus and getting result",
-                iota_metrics::SUBSECOND_LATENCY_SEC_BUCKETS.to_vec(),
-                registry,
-            )
-            .unwrap(),
-            handle_transaction_latency: register_histogram_with_registry!(
-                "validator_service_handle_transaction_latency",
-                "Latency of handling a transaction",
-                iota_metrics::SUBSECOND_LATENCY_SEC_BUCKETS.to_vec(),
-                registry,
-            )
-            .unwrap(),
-            handle_certificate_consensus_latency: register_histogram_with_registry!(
-                "validator_service_handle_certificate_consensus_latency",
-                "Latency of handling a consensus transaction certificate",
-                iota_metrics::COARSE_LATENCY_SEC_BUCKETS.to_vec(),
-                registry,
-            )
-            .unwrap(),
-            submit_certificate_consensus_latency: register_histogram_with_registry!(
-                "validator_service_submit_certificate_consensus_latency",
-                "Latency of submit_certificate RPC handler",
-                iota_metrics::COARSE_LATENCY_SEC_BUCKETS.to_vec(),
-                registry,
-            )
-            .unwrap(),
-            handle_certificate_non_consensus_latency: register_histogram_with_registry!(
-                "validator_service_handle_certificate_non_consensus_latency",
-                "Latency of handling a non-consensus transaction certificate",
-                iota_metrics::SUBSECOND_LATENCY_SEC_BUCKETS.to_vec(),
-                registry,
-            )
-            .unwrap(),
-            handle_soft_bundle_certificates_consensus_latency: register_histogram_with_registry!(
-                "validator_service_handle_soft_bundle_certificates_consensus_latency",
-                "Latency of handling a consensus soft bundle",
-                iota_metrics::COARSE_LATENCY_SEC_BUCKETS.to_vec(),
-                registry,
-            )
-            .unwrap(),
-            handle_soft_bundle_certificates_count: register_histogram_with_registry!(
-                "validator_service_handle_soft_bundle_certificates_count",
-                "The number of certificates included in a soft bundle",
-                iota_metrics::COUNT_BUCKETS.to_vec(),
-                registry,
-            )
-            .unwrap(),
-            handle_soft_bundle_certificates_size_bytes: register_histogram_with_registry!(
-                "validator_service_handle_soft_bundle_certificates_size_bytes",
-                "The size of soft bundle in bytes",
-                iota_metrics::BYTES_BUCKETS.to_vec(),
-                registry,
-            )
-            .unwrap(),
-            handle_capability_notification_latency: register_histogram_with_registry!(
-                "validator_service_handle_capability_notification_latency",
-                "Latency of handling a capability notification",
-                iota_metrics::SUBSECOND_LATENCY_SEC_BUCKETS.to_vec(),
-                registry,
-            )
-            .unwrap(),
-            num_rejected_tx_in_epoch_boundary: register_int_counter_with_registry!(
-                "validator_service_num_rejected_tx_in_epoch_boundary",
-                "Number of rejected transaction during epoch transitioning",
-                registry,
-            )
-            .unwrap(),
-            num_rejected_cert_in_epoch_boundary: register_int_counter_with_registry!(
-                "validator_service_num_rejected_cert_in_epoch_boundary",
-                "Number of rejected transaction certificate during epoch transitioning",
-                registry,
-            )
-            .unwrap(),
-            num_rejected_tx_during_overload: register_int_counter_vec_with_registry!(
-                "validator_service_num_rejected_tx_during_overload",
-                "Number of rejected transaction due to system overload",
-                &["error_type"],
-                registry,
-            )
-            .unwrap(),
-            num_rejected_cert_during_overload: register_int_counter_vec_with_registry!(
-                "validator_service_num_rejected_cert_during_overload",
-                "Number of rejected transaction certificate due to system overload",
-                &["error_type"],
-                registry,
-            )
-            .unwrap(),
-            num_rejected_capability_notifications_during_overload: register_int_counter_vec_with_registry!(
-                "num_rejected_capability_notifications_during_overload",
-                "Number of rejected capability notifications from non-committee active validators due to system overload",
-                &["error_type"],
-                registry,
-            )
-            .unwrap(),
-            connection_ip_not_found: register_int_counter_with_registry!(
-                "validator_service_connection_ip_not_found",
-                "Number of times connection IP was not extractable from request",
-                registry,
-            )
-            .unwrap(),
-            forwarded_header_parse_error: register_int_counter_with_registry!(
-                "validator_service_forwarded_header_parse_error",
-                "Number of times x-forwarded-for header could not be parsed",
-                registry,
-            )
-            .unwrap(),
-            forwarded_header_invalid: register_int_counter_with_registry!(
-                "validator_service_forwarded_header_invalid",
-                "Number of times x-forwarded-for header was invalid",
-                registry,
-            )
-            .unwrap(),
-            forwarded_header_not_included: register_int_counter_with_registry!(
-                "validator_service_forwarded_header_not_included",
-                "Number of times x-forwarded-for header was (unexpectedly) not included in request",
-                registry,
-            )
-            .unwrap(),
-            client_id_source_config_mismatch: register_int_counter_with_registry!(
-                "validator_service_client_id_source_config_mismatch",
-                "Number of times detected that client id source config doesn't agree with x-forwarded-for header",
-                registry,
-            )
-            .unwrap(),
-        }
-    }
-
-    /// Creates a new `ValidatorServiceMetrics` for testing.
-    pub fn new_for_tests() -> Self {
-        let registry = Registry::new();
-        Self::new(&registry)
-    }
-}
-
-/// The validator service.
-#[derive(Clone)]
-pub struct ValidatorService {
-    state: Arc<AuthorityState>,
-    consensus_adapter: Arc<ConsensusAdapter>,
-    metrics: Arc<ValidatorServiceMetrics>,
-    traffic_controller: Option<Arc<TrafficController>>,
-    client_id_source: Option<ClientIdSource>,
-}
-
 impl ValidatorService {
-    /// Creates a new `ValidatorService`.
-    pub fn new(
-        state: Arc<AuthorityState>,
-        consensus_adapter: Arc<ConsensusAdapter>,
-        validator_metrics: Arc<ValidatorServiceMetrics>,
-        traffic_controller_metrics: TrafficControllerMetrics,
-        policy_config: Option<PolicyConfig>,
-        firewall_config: Option<RemoteFirewallConfig>,
-    ) -> Self {
-        Self {
-            state,
-            consensus_adapter,
-            metrics: validator_metrics,
-            traffic_controller: policy_config.clone().map(|policy| {
-                Arc::new(TrafficController::init(
-                    policy,
-                    traffic_controller_metrics,
-                    firewall_config,
-                ))
-            }),
-            client_id_source: policy_config.map(|policy| policy.client_id_source),
-        }
-    }
-
-    pub fn new_for_tests(
-        state: Arc<AuthorityState>,
-        consensus_adapter: Arc<ConsensusAdapter>,
-        metrics: Arc<ValidatorServiceMetrics>,
-    ) -> Self {
-        Self {
-            state,
-            consensus_adapter,
-            metrics,
-            traffic_controller: None,
-            client_id_source: None,
-        }
-    }
-
-    /// Returns the validator state.
-    pub fn validator_state(&self) -> &Arc<AuthorityState> {
-        &self.state
-    }
-
     /// Executes a `CertifiedTransaction` for testing.
     pub async fn execute_certificate_for_testing(
         &self,
@@ -729,11 +364,7 @@ impl ValidatorService {
 
         Ok((Some(responses), Weight::zero()))
     }
-}
 
-type WrappedServiceResponse<T> = Result<(tonic::Response<T>, Weight), tonic::Status>;
-
-impl ValidatorService {
     async fn transaction_impl(
         &self,
         request: tonic::Request<Transaction>,
@@ -1042,142 +673,6 @@ impl ValidatorService {
         Ok((tonic::Response::new(response), Weight::one()))
     }
 
-    fn get_client_ip_addr<T>(
-        &self,
-        request: &tonic::Request<T>,
-        source: &ClientIdSource,
-    ) -> Option<IpAddr> {
-        match source {
-            ClientIdSource::SocketAddr => {
-                let socket_addr: Option<SocketAddr> = request.remote_addr();
-
-                // We will hit this case if the IO type used does not
-                // implement Connected or when using a unix domain socket.
-                // TODO: once we have confirmed that no legitimate traffic
-                // is hitting this case, we should reject such requests that
-                // hit this case.
-                if let Some(socket_addr) = socket_addr {
-                    Some(socket_addr.ip())
-                } else {
-                    if cfg!(msim) {
-                        // Ignore the error from simtests.
-                    } else if cfg!(test) {
-                        panic!("Failed to get remote address from request");
-                    } else {
-                        self.metrics.connection_ip_not_found.inc();
-                        error!("Failed to get remote address from request");
-                    }
-                    None
-                }
-            }
-            ClientIdSource::XForwardedFor(num_hops) => {
-                let do_header_parse = |op: &MetadataValue<Ascii>| {
-                    match op.to_str() {
-                        Ok(header_val) => {
-                            let header_contents =
-                                header_val.split(',').map(str::trim).collect::<Vec<_>>();
-                            if *num_hops == 0 {
-                                error!(
-                                    "x-forwarded-for: 0 specified. x-forwarded-for contents: {:?}. Please assign nonzero value for \
-                                    number of hops here, or use `socket-addr` client-id-source type if requests are not being proxied \
-                                    to this node. Skipping traffic controller request handling.",
-                                    header_contents,
-                                );
-                                return None;
-                            }
-                            let contents_len = header_contents.len();
-                            if contents_len < *num_hops {
-                                error!(
-                                    "x-forwarded-for header value of {:?} contains {} values, but {} hops were specified. \
-                                    Expected at least {} values. Please correctly set the `x-forwarded-for` value under \
-                                    `client-id-source` in the node config.",
-                                    header_contents, contents_len, num_hops, contents_len,
-                                );
-                                self.metrics.client_id_source_config_mismatch.inc();
-                                return None;
-                            }
-                            let Some(client_ip) = header_contents.get(contents_len - num_hops)
-                            else {
-                                error!(
-                                    "x-forwarded-for header value of {:?} contains {} values, but {} hops were specified. \
-                                    Expected at least {} values. Skipping traffic controller request handling.",
-                                    header_contents, contents_len, num_hops, contents_len,
-                                );
-                                return None;
-                            };
-                            parse_ip(client_ip).or_else(|| {
-                                self.metrics.forwarded_header_parse_error.inc();
-                                None
-                            })
-                        }
-                        Err(e) => {
-                            // TODO: once we have confirmed that no legitimate traffic
-                            // is hitting this case, we should reject such requests that
-                            // hit this case.
-                            self.metrics.forwarded_header_invalid.inc();
-                            error!("Invalid UTF-8 in x-forwarded-for header: {:?}", e);
-                            None
-                        }
-                    }
-                };
-                if let Some(op) = request.metadata().get("x-forwarded-for") {
-                    do_header_parse(op)
-                } else if let Some(op) = request.metadata().get("X-Forwarded-For") {
-                    do_header_parse(op)
-                } else {
-                    self.metrics.forwarded_header_not_included.inc();
-                    error!(
-                        "x-forwarded-for header not present for request despite node configuring x-forwarded-for tracking type"
-                    );
-                    None
-                }
-            }
-        }
-    }
-
-    async fn handle_traffic_req(&self, client: Option<IpAddr>) -> Result<(), tonic::Status> {
-        if let Some(traffic_controller) = &self.traffic_controller {
-            if !traffic_controller.check(&client, &None).await {
-                // Entity in blocklist
-                Err(tonic::Status::from_error(IotaError::TooManyRequests.into()))
-            } else {
-                Ok(())
-            }
-        } else {
-            Ok(())
-        }
-    }
-
-    fn handle_traffic_resp<T>(
-        &self,
-        client: Option<IpAddr>,
-        wrapped_response: WrappedServiceResponse<T>,
-    ) -> Result<tonic::Response<T>, tonic::Status> {
-        let (error, spam_weight, unwrapped_response) = match wrapped_response {
-            Ok((result, spam_weight)) => (None, spam_weight, Ok(result)),
-            Err(status) => (
-                Some(IotaError::from(status.clone())),
-                Weight::zero(),
-                Err(status),
-            ),
-        };
-
-        if let Some(traffic_controller) = self.traffic_controller.clone() {
-            traffic_controller.tally(TrafficTally {
-                direct: client,
-                through_fullnode: None,
-                error_info: error.map(|e| {
-                    let error_type = String::from(e.as_ref());
-                    let error_weight = normalize(e);
-                    (error_weight, error_type)
-                }),
-                spam_weight,
-                timestamp: SystemTime::now(),
-            })
-        }
-        unwrapped_response
-    }
-
     async fn handle_capability_notification_v1_impl(
         &self,
         request: tonic::Request<HandleCapabilityNotificationRequestV1>,
@@ -1194,8 +689,8 @@ impl ValidatorService {
             }
             .into()
         );
-        // Validate if cert can be executed
-        // Fullnode does not serve handle_certificate call.
+        // Validate if the capability notification can be handled.
+        // Fullnode does not serve capability notification requests.
         fp_ensure!(
             !self.state.is_fullnode(&epoch_store),
             IotaError::FullNodeCantHandleAuthorityCapabilities.into()
@@ -1581,7 +1076,7 @@ impl ValidatorService {
                 }
             },
         )
-        .await;
+            .await;
 
         // The select! produces Either<Vec<EffectsDigest>, IotaError>.
         // Unpack the three outcomes: executed, dropped, or timeout.
@@ -1673,55 +1168,6 @@ impl ValidatorService {
     }
 }
 
-fn make_tonic_request_for_testing<T>(message: T) -> tonic::Request<T> {
-    // simulate a TCP connection, which would have added extensions to
-    // the request object that would be used downstream
-    let mut request = tonic::Request::new(message);
-    let tcp_connect_info = TcpConnectInfo {
-        local_addr: None,
-        remote_addr: Some(SocketAddr::new([127, 0, 0, 1].into(), 0)),
-    };
-    request.extensions_mut().insert(tcp_connect_info);
-    request
-}
-
-// TODO: refine error matching here
-fn normalize(err: IotaError) -> Weight {
-    match err {
-        IotaError::UserInput {
-            error: UserInputError::IncorrectUserSignature { .. },
-        } => Weight::one(),
-        IotaError::InvalidSignature { .. }
-        | IotaError::SignerSignatureAbsent { .. }
-        | IotaError::SignerSignatureNumberMismatch { .. }
-        | IotaError::IncorrectSigner { .. }
-        | IotaError::UnknownSigner { .. }
-        | IotaError::WrongEpoch { .. } => Weight::one(),
-        _ => Weight::zero(),
-    }
-}
-
-/// Implements generic pre- and post-processing. Since this is on the critical
-/// path, any heavy lifting should be done in a separate non-blocking task
-/// unless it is necessary to override the return value.
-#[macro_export]
-macro_rules! handle_with_decoration {
-    ($self:ident, $func_name:ident, $request:ident) => {{
-        if $self.client_id_source.is_none() {
-            return $self.$func_name($request).await.map(|(result, _)| result);
-        }
-
-        let client = $self.get_client_ip_addr(&$request, $self.client_id_source.as_ref().unwrap());
-
-        // check if either IP is blocked, in which case return early
-        $self.handle_traffic_req(client.clone()).await?;
-
-        // handle traffic tallying
-        let wrapped_response = $self.$func_name($request).await;
-        $self.handle_traffic_resp(client, wrapped_response)
-    }};
-}
-
 #[async_trait]
 impl Validator for ValidatorService {
     /// Handles a `Transaction` request.
@@ -1744,6 +1190,20 @@ impl Validator for ValidatorService {
         .unwrap()
     }
 
+    async fn handle_certificate_v1(
+        &self,
+        request: tonic::Request<HandleCertificateRequestV1>,
+    ) -> Result<tonic::Response<HandleCertificateResponseV1>, tonic::Status> {
+        handle_with_decoration!(self, handle_certificate_v1_impl, request)
+    }
+
+    async fn handle_soft_bundle_certificates_v1(
+        &self,
+        request: tonic::Request<HandleSoftBundleCertificatesRequestV1>,
+    ) -> Result<tonic::Response<HandleSoftBundleCertificatesResponseV1>, tonic::Status> {
+        handle_with_decoration!(self, handle_soft_bundle_certificates_v1_impl, request)
+    }
+
     /// Submits a `CertifiedTransaction` request.
     async fn submit_certificate(
         &self,
@@ -1762,20 +1222,6 @@ impl Validator for ValidatorService {
         })
         .await
         .unwrap()
-    }
-
-    async fn handle_certificate_v1(
-        &self,
-        request: tonic::Request<HandleCertificateRequestV1>,
-    ) -> Result<tonic::Response<HandleCertificateResponseV1>, tonic::Status> {
-        handle_with_decoration!(self, handle_certificate_v1_impl, request)
-    }
-
-    async fn handle_soft_bundle_certificates_v1(
-        &self,
-        request: tonic::Request<HandleSoftBundleCertificatesRequestV1>,
-    ) -> Result<tonic::Response<HandleSoftBundleCertificatesResponseV1>, tonic::Status> {
-        handle_with_decoration!(self, handle_soft_bundle_certificates_v1_impl, request)
     }
 
     /// Handles an `ObjectInfoRequest` request.
