+++ crates/iota-core/src/unit_tests/transaction_tests.rs
@@ -43,6 +43,7 @@ use crate::{
         authority_tests::{call_move_, create_gas_objects, publish_object_basics},
         test_authority_builder::TestAuthorityBuilder,
     },
+    authority_client::validator::ValidatorAPI,
     consensus_adapter::consensus_tests::make_consensus_adapter_for_test,
     mock_consensus::with_block_status,
 };
@@ -70,7 +71,7 @@ use iota_types::{
 use super::*;
 pub use crate::authority::authority_test_utils::init_state_with_ids;
 use crate::{
-    authority_client::{AuthorityAPI, NetworkAuthorityClient},
+    authority_client::NetworkAuthorityClient,
     authority_server::{AuthorityServer, AuthorityServerHandle},
     stake_aggregator::{InsertResult, StakeAggregator},
 };
