+++ crates/iota-tool/src/commands.rs
@@ -14,7 +14,9 @@ use iota_config::{
     genesis::Genesis,
     object_storage_config::{ObjectStoreConfig, ObjectStoreType},
 };
-use iota_core::{authority_aggregator::AuthorityAggregatorBuilder, authority_client::AuthorityAPI};
+use iota_core::{
+    authority_aggregator::AuthorityAggregatorBuilder, authority_client::validator::ValidatorAPI,
+};
 use iota_protocol_config::Chain;
 use iota_replay::{ReplayToolCommand, execute_replay_command};
 use iota_sdk::{IotaClient, IotaClientBuilder, rpc_types::IotaTransactionBlockResponseOptions};
